fix: address final whole-branch review findings
- PackAggregator: widen SoC and capacity to Double BEFORE multiplying in
  the parallel capacity-weighted average. The Float x Float product
  rounded before widening, so a single pack with soc=19f, capacity=13.6f
  aggregated to 18.999998f — truncating to 18% in the notification and
  tripping the low-SoC alert a point early.
- PackAggregatorTest: add a hostile-fixture identity test (19f x 13.6f,
  no tolerance) that fails on the old code and asserts soc.toInt()
  survives aggregation. Verified RED before the fix.
- KableBmsRepository: correct the vehicleConnection KDoc to name BOTH
  cleanup helpers and their opposite locking contracts (comment only).
- connectDemo: also reset _activeVehicleData inside the existing locked
  section, so a previous real connection's packs are not left published
  in demo mode (matches disconnect()).
- PackRow.sq: delete the dead deleteFromIndex query left over from the
  replaced index-based trim; verified it has no callers.

+++ composeApp/src/commonMain/kotlin/ru/sodovaya/volty/domain/stats/PackAggregator.kt
@@ -60,8 +60,12 @@ object PackAggregator {
                 // Capacity-weighted average of the packs' reported SoC —
                 // reduces to the identity for a single pack. Falls back to
                 // the plain mean when no pack reports capacity.
+                // Widen BEFORE multiplying: a Float x Float product rounds to
+                // Float precision first, breaking the single-pack identity
+                // (e.g. 19f x 13.6f aggregates to 18.999998f, truncating to
+                // 18% and tripping the low-SoC alert a point early).
                 soc = if (capacity > 0f)
-                    (data.sumOf { (it.soc * it.capacity).toDouble() } / capacity).toFloat()
+                    (data.sumOf { it.soc.toDouble() * it.capacity.toDouble() } / capacity).toFloat()
                 else data.map { it.soc }.average().toFloat()
             }
             PackTopology.SERIES -> {
